Move all article view related struct to ArticleView file
