+++ NewsAppPlus/Views/NewsView.swift
@@ -22,7 +22,7 @@ struct NewsView: View {
                 TabView(selection: $currentSlideIndex) {
                     ForEach(0..<min(newsViewModel.articles.filter { $0.source.name == selectedSource }.count, maxSlides), id: \.self) { slide in
                         
-                        ArticleDetails(newArticle: newsViewModel.articles.filter { $0.source.name == selectedSource }[slide])
+                        ArticleNewsDetails(newArticle: newsViewModel.articles.filter { $0.source.name == selectedSource }[slide])
                     }
                 }
                 .tabViewStyle(PageTabViewStyle())
@@ -40,7 +40,7 @@ struct NewsView: View {
                 List(newsViewModel.articles.filter { $0.source.name == selectedSource }, id: \.url) { article in
                     
                     // Display article details
-                    ArticleDetails(newArticle: article)
+                    ArticleNewsDetails(newArticle: article)
                     
                 }
                 .listStyle(.plain)
@@ -62,74 +62,5 @@ struct NewsView: View {
                 currentSlideIndex = (currentSlideIndex + 1) % min(newsViewModel.articles.filter { $0.source.name == selectedSource }.count, maxSlides)
             }
         }
-    }
-}
-
-private struct ArticleDetails: View {
-    var newArticle: ArticleDto
-    
-    var body: some View {
-        VStack(alignment: .leading) {
-            // Display article image
-            let imageUrl = newArticle.map().urlToImage
-            ArticleImage(newImageUrl: imageUrl, newArticleUrl: newArticle.map().url)
-            
-            // Display article content
-            ArticleContent(newArticle: newArticle)
-        }
-    }
-}
-
-private struct ArticleImage: View {
-    var newImageUrl: String
-    var newArticleUrl: String
-    
-    var body: some View {
-        ZStack {
-            NavigationLink(destination: NewsWebView(urlString: newArticleUrl)) { EmptyView() }
-            AsyncImage(url: URL(string: newImageUrl)) { image in
-                image.resizable()
-                    .modifier(ImageModifier())
-            }
-        placeholder: {
-            
-            Image(systemName: Constant.IMAGE_PLACEHOLDER)
-                .resizable()
-                .modifier(ImageModifier())
-        }
-            
-        }
-    }
-}
-
-private struct ArticleContent: View {
-    var newArticle: ArticleDto
-    
-    var body: some View {
-        Text(newArticle.map().title)
-            .font(.headline)
-            .fontWeight(.black)
-        
-        Text(newArticle.map().description)
-            .font(.subheadline)
-            .lineLimit(2)
-        
-        Text(newArticle.source.name)
-            .font(.footnote)
-            .foregroundStyle(.pink)
-        
-        Text(newArticle.map().author)
-            .font(.footnote)
-            .bold()
-            .foregroundStyle(.gray)
-    }
-}
-
-struct ImageModifier: ViewModifier {
-    func body(content: Content) -> some View {
-        content
-            .aspectRatio(contentMode: .fill)
-            .frame(width: UIScreen.main.bounds.width * 0.95/1, height: UIScreen.main.bounds.height * 1/3.5)
-            .cornerRadius(12)
     }
 }
